⚙️ refactor: update import path for AICommandsManager to the new directory structure
